+++ packages/algob/src/lib/credentials.ts
@@ -37,15 +37,20 @@ export function KMDCredentialsFromEnv(): NetworkCredentials {
 	const kmdAddr = process.env.KMD_ADDR;
 
 	if (token === undefined && kmdAddr === undefined) {
-		const kmdData = process.env.$KMD_DATA;
+		const kmdData = process.env.KMD_DATA;
 		if (kmdData === undefined) {
-			throw new Error("Either KMD Credentials or $KMD_DATA should be defined in env");
+			throw new Error("Either KMD Credentials or KMD_DATA should be defined in env");
 		}
 
 		const loadToken = fs.readFileSync(path.join(kmdData, "kmd.token"), "utf8");
 		const loadNet = fs.readFileSync(path.join(kmdData, "kmd.net"), "utf8");
 		const arr = loadNet.toString().split(":");
-
+		if (arr.length !== 2) {
+			throw new Error("Wrong network address in kmd.net file. Expected: [http://]host:port");
+		}
+		if (!arr[0].startsWith("http")) {
+			arr[0] = "http://" + arr[0];
+		}
 		return {
 			host: arr[0],
 			port: +arr[1],